move createRelayNode under an explicit export
