+++ packages/sdk/src/create.ts
@@ -22,7 +22,6 @@ import type {
   Libp2pComponents,
   LightNode,
   ProtocolCreateOptions,
-  RelayNode,
   ShardingParams
 } from "@waku/interfaces";
 import { wakuPeerExchangeDiscovery } from "@waku/peer-exchange";
@@ -75,48 +74,6 @@ export async function createLightNode(
     lightPush,
     filter
   ) as LightNode;
-}
-
-/**
- * Create a Waku node that uses Waku Relay to send and receive messages,
- * enabling some privacy preserving properties.
- * * @remarks
- * This function creates a Relay Node using the Waku Relay protocol.
- * While it is technically possible to use this function in a browser environment,
- * it is not recommended due to potential performance issues and limited browser capabilities.
- * If you are developing a browser-based application, consider alternative approaches like creating a Light Node
- * or use this function with caution.
- */
-export async function createRelayNode(
-  options?: ProtocolCreateOptions & WakuOptions & Partial<RelayCreateOptions>
-): Promise<RelayNode> {
-  options = options ?? {};
-
-  const libp2pOptions = options?.libp2p ?? {};
-  const peerDiscovery = libp2pOptions.peerDiscovery ?? [];
-  if (options?.defaultBootstrap) {
-    peerDiscovery.push(...defaultPeerDiscoveries());
-    Object.assign(libp2pOptions, { peerDiscovery });
-  }
-
-  const libp2p = await defaultLibp2p(
-    options.shardInfo,
-    wakuGossipSub(options),
-    libp2pOptions,
-    options?.userAgent
-  );
-
-  const relay = wakuRelay(options);
-
-  return new WakuNode(
-    options,
-    libp2p,
-    options.shardInfo,
-    undefined,
-    undefined,
-    undefined,
-    relay
-  ) as RelayNode;
 }
 
 /**
