+++ packages/tests/tests/relay/publish.node.spec.ts
@@ -1,6 +1,6 @@
 import { createEncoder } from "@waku/core";
 import { IRateLimitProof, RelayNode, SendError } from "@waku/interfaces";
-import { createRelayNode } from "@waku/sdk";
+import { createRelayNode } from "@waku/sdk/relay";
 import { utf8ToBytes } from "@waku/utils/bytes";
 import { expect } from "chai";
